\docType{package}
\name{DirectedClustering-package}
\alias{DirectedClustering}
\alias{DirectedClustering-package}
\title{
Alternative Clustering Coefficients for Directed/Undirected and Weighted Networks
}
\description{
The \pkg{DirectedClustering} \R package presented here includes an enhanced \R implementation of Local and Global (average) Clustering Coefficients for Directed/Undirected and Unweighted/Weighted Networks. \cr Functions are based on Barrat et al. (2004) and Onnela et al. (2005) coefficients when the network is undirected, while it is based on Fagiolo (2007) and Clemente and Grassi (2018) proposals when the network is directed. In the directed case, different components of directed clustering coefficient are also considered.

Details of alternative coefficients computed by \pkg{DirectedClustering} \R package can be found in Clemente, Grassi (2018).

Please report any issue arising or bug in the code to gianpaolo.clemente@unicatt.it.
}
\author{
Gian Paolo Clemente, gianpaolo.clemente@unicatt.it, Rosanna Grassi, rosanna.grassi@unimib.it}
\references{
Clemente, G.P. and Grassi, R. (2018) Directed clustering in weighted networks: a new perspective, \emph{Chaos, Solitons and Fractals}, \strong{107},26--38.
}
\note{
This package and the functions herein are provided as is, without any guarantee regarding the accuracy of calculations. The authors disclaim any liability arising by any losses due to direct and indirect use of this package.
}
